Various changes to osspley/ossrecord

+++ cmd/ossplay/ossplay_wparser.h
@@ -3,8 +3,8 @@
 
 #include "ossplay.h"
 
-int write_head (FILE *, fctypes_t, unsigned int, int, int, int);
-int finalize_head (FILE *, fctypes_t, unsigned int, int, int, int);
+int write_head (FILE *, fctypes_t, unsigned long long, int, int, int);
+int finalize_head (FILE *, fctypes_t, unsigned long long, int, int, int);
 
 #ifdef OSS_BIG_ENDIAN
 #define BE_INT(x) x
